backend/requirements.txt
Flask
flask-cors
flask-sqlalchemy
Werkzeug
pyjwt==2.1.0
python-docx
requests
beautifulsoup4
pydantic
pandas
numpy
matplotlib
statsmodels
babel
openpyxl
seaborn
reportlab
python-docx
reportlab
markdown
markdown2
